+++ src/core/pluginAdapter.js
@@ -1,7 +1,7 @@
 // @flow
 
-import type {Graph, NodeAddressT, EdgeAddressT} from "../core/graph";
-import type {Repo} from "../core/repo";
+import type {Graph, NodeAddressT, EdgeAddressT} from "./graph";
+import type {Repo} from "./repo";
 
 export type EdgeType = {|
   +forwardName: string,
